backend/requirements.txt
fastapi
uvicorn
PyPDF2
python-dotenv
google-generativeai

